# external-svc/openapi.yaml
openapi: 3.0.1
info:
  title: Choreo BYOC Examples
  description: This repository contains demo API implemented in different frameworks to showcase capabilities of choreo.
  version: '0.1'
paths:
  /proxy/:
    post:
      description: Auto generated using Swagger Inspector
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                args:
                  type: string
                host:
                  type: string
            examples:
              '0':
                value: |-
                  {
                      "host": "http://postman-echo.com",
                      "args": "get?foo1=bar1&foo2=bar2"
                  }
      responses:
        '200':
          description: Auto generated using Swagger Inspector
          content:
            text/html; charset=utf-8:
              schema:
                type: string
              examples: { }
  /print-body:
    post:
      description: Receives and prints the request body to server logs
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: Any JSON object
          text/plain:
            schema:
              type: string
              description: Any plain text content
          application/xml:
            schema:
              type: string
              description: Any XML content
        description: Request body content to be printed to logs
      responses:
        '200':
          description: Confirmation that body was received and printed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  body_length:
                    type: integer
              examples:
                '0':
                  value: |
                    {"message": "Body received and printed", "body_length": 35}
        '405':
          description: Method not allowed - only POST is supported
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                '0':
                  value: |
                    {"error": "Method not allowed. Use POST."}
        '500':
          description: Internal server error reading request body
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                '0':
                  value: |
                    {"error": "Failed to read request body"}
  /print-headers:
    get:
      description: Receives and prints all request headers to server logs in a single line. Send any custom headers and they will be printed.
      parameters:
        - name: X-Custom-Header
          in: header
          required: false
          schema:
            type: string
          description: Example custom header that will be printed to logs
          example: custom-value
      responses:
        '200':
          description: Confirmation that headers were received and printed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  header_count:
                    type: integer
              examples:
                '0':
                  value: |
                    {"message": "Headers received and printed", "header_count": 5}
        '405':
          description: Method not allowed - only GET is supported
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                '0':
                  value: |
                    {"error": "Method not allowed. Use GET."}
  /healthz/:
    get:
      description: Auto generated using Swagger Inspector
      responses:
        '200':
          description: Auto generated using Swagger Inspector
          content:
            application/json:
              schema:
                type: object
                properties:
                  healthy:
                    type: boolean
              examples:
                '0':
                  value: |
                    {"healthy":true}
  /:
    get:
      description: Auto generated using Swagger Inspector
      responses:
        '200':
          description: Auto generated using Swagger Inspector
          content:
            application/json:
              schema:
                type: object
                properties:
                  healthy:
                    type: boolean
              examples:
                '0':
                  value: |
                    {"active":true}
  /hello/:
    get:
      description: Auto generated using Swagger Inspector
      parameters:
        - name: name
          in: query
          schema:
            type: string
          example: world
      responses:
        '200':
          description: Auto generated using Swagger Inspector
          content:
            text/html; charset=utf-8:
              schema:
                type: string
              examples: { }
  /five:
    get:
      description: get 500
      responses:
        '500':
          description: simulating 500
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
              examples:
                '0':
                  value: |
                    {"status":500}
  /four09:
    get:
      description: get 409
      responses:
        '500':
          description: simulating 409
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
              examples:
                '0':
                  value: |
                    {"status":409}
  /call-internal:
    get:
      description: Calls the internal service's greeting endpoint and returns the response
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: Name to pass to the internal service greeting
          example: world
      responses:
        '200':
          description: Response from the internal service
          content:
            application/json:
              schema:
                type: object
                properties:
                  greeting:
                    type: string
                  timestamp:
                    type: string
        '502':
          description: Failed to reach internal service
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string